refactor: split API endpoints and fix compose builds
