+++ ops/docker-compose.yml
@@ -13,7 +13,9 @@ services:
       - pgdata:/var/lib/postgresql/data
       - ./init.sql:/docker-entrypoint-initdb.d/init.sql
   api:
-    build: ../backend/DesignDocs.Api
+    build:
+      context: ../backend/DesignDocs.Api
+      dockerfile: Dockerfile
     environment:
       ASPNETCORE_URLS: ${ASPNETCORE_URLS}
       DB_CONNECTION: ${DB_CONNECTION}
@@ -25,7 +27,9 @@ services:
     depends_on:
       - postgres
   web:
-    build: ../frontend
+    build:
+      context: ../frontend
+      dockerfile: Dockerfile
     environment:
       VITE_API_URL: http://api:8090
     ports:
